Complete blogilista for osa 44.

+++ osa4/blogilista/controllers/blogs.js
@@ -1,11 +1,17 @@
+const jwt = require("jsonwebtoken");
+
 const router = require("express").Router();
 module.exports = router;
 
 const Blog = require("../models/blogs");
+const User = require("../models/users");
 
 router.get("/", async (request, response, next) => {
   try {
-    const blogs = await Blog.find({});
+    const blogs = await Blog.find({}).populate("user", {
+      username: 1,
+      name: 1,
+    });
     response.json(blogs);
   } catch (exception) {
     next(exception);
@@ -17,15 +23,26 @@ router.post("/", async (request, response) => {
     response.status(400).end();
     return;
   }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
+
   const blog = new Blog({
     title: request.body.title,
     author: request.body.author,
+    user: user,
     url: request.body.url,
     likes: request.body.likes || 0,
   });
 
-  const result = await blog.save();
-  response.status(201).json(result);
+  const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
+  response.status(201).json(savedBlog);
 });
 
 router.get("/:id", async (request, response) => {
@@ -42,6 +59,10 @@ router.put("/:id", async (request, response) => {
     response.status(400).end();
     return;
   }
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
   const blog = {
     title: request.body.title,
     author: request.body.author,
@@ -57,6 +78,22 @@ router.put("/:id", async (request, response) => {
 });
 
 router.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+  const blog = await Blog.findById(request.params.id);
+  if (blog) {
+    if (blog.user.toString() === decodedToken.id) {
+      await Blog.findByIdAndRemove(request.params.id);
+      response.status(204).end();
+    } else {
+      response
+        .status(400)
+        .send({ error: "Blogs can only be deleted by the user who added them" })
+        .end();
+    }
+  } else {
+    response.status(204).end();
+  }
 });
